+++ src/main/java/pl/coderslab/model/service/category/CategoryRepository.java
@@ -3,5 +3,5 @@ package pl.coderslab.model.service.category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.coderslab.model.entity.Category;
 
-interface CategoryRepository extends JpaRepository<Category, Double> {
+interface CategoryRepository extends JpaRepository<Category, Long> {
 }
