+++ src/main/java/pl/coderslab/model/service/donation/DonationRepository.java
@@ -3,5 +3,5 @@ package pl.coderslab.model.service.donation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.coderslab.model.entity.Donation;
 
-public interface DonationRepository extends JpaRepository<Donation, Double> {
+public interface DonationRepository extends JpaRepository<Donation, Long> {
 }
