+++ src/main/java/pl/coderslab/model/service/institution/InstitutionRepository.java
@@ -3,5 +3,5 @@ package pl.coderslab.model.service.institution;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.coderslab.model.entity.Institution;
 
-interface InstitutionRepository extends JpaRepository<Institution, Double> {
+interface InstitutionRepository extends JpaRepository<Institution, Long> {
 }
